requirements.txt
py-clob-client>=0.24.0
python-dotenv==1.1.1
pandas
gspread
gspread-dataframe
sortedcontainers
black==24.4.2
eth-account>=0.13.0
eth-utils>=5.0.0
poly_eip712_structs==0.0.1
py_order_utils>=0.3.2
pytest==8.2.2
requests==2.32.3
websockets==12.0
cryptography==42.0.8
google-auth
setuptools
web3>=7.6.0
fastapi==0.115.4
uvicorn[standard]==0.30.6
trafilatura
python-dateutil
tqdm
openai>=1.37.0
httpx>=0.27.0
tenacity>=8.4.1
beautifulsoup4>=4.12.3
html2text>=2024.2.26
curl_cffi
authlib
prometheus_client>=0.20.0